Masih belom..

+++ src/main/java/com/apap/tugas1/repository/InstansiDB.java
@@ -9,5 +9,5 @@ import com.apap.tugas1.model.InstansiModel;
 
 @Repository
 public interface InstansiDB extends JpaRepository<InstansiModel, Long> {
-	InstansiModel findById(BigInteger id);
+	InstansiModel findById(long id);
 }
