+++ src/main/java/com/apap/tugas1/repository/ProvinsiDB.java
@@ -9,5 +9,5 @@ import com.apap.tugas1.model.ProvinsiModel;
 
 @Repository
 public interface ProvinsiDB extends JpaRepository<ProvinsiModel, Long> {
-	ProvinsiModel findById(BigInteger id);
+	ProvinsiModel findById(long id);
 }
