+++ src/main/java/com/apap/tugas1/service/JabatanService.java
@@ -6,8 +6,8 @@ import java.util.List;
 import com.apap.tugas1.model.JabatanModel;
 
 public interface JabatanService {
-		JabatanModel getJabatanDetailById(BigInteger id);
-		List<JabatanModel> getJabatanList();
-		void addJabatan(JabatanModel jabatan);
-		void deleteJabatan(BigInteger id);
+	JabatanModel getJabatanDetailById(long id);
+	List<JabatanModel> getJabatanList();
+	void addJabatan(JabatanModel jabatan);
+	void deleteJabatan(long id);
 }
