+++ src/main/java/com/apap/tugas1/service/ProvinsiService.java
@@ -3,13 +3,12 @@ package com.apap.tugas1.service;
 import java.math.BigInteger;
 import java.util.List;
 
-import com.apap.tugas1.model.InstansiModel;
 import com.apap.tugas1.model.PegawaiModel;
 import com.apap.tugas1.model.ProvinsiModel;
 
 public interface ProvinsiService {
 	List<ProvinsiModel> getProvinsiList();
-	ProvinsiModel getProvinsiDetailById(BigInteger id);
-	List<PegawaiModel> getPegawaiInInstansiInProvinsiById(BigInteger idProvinsi, BigInteger idInstansi);
-	List<PegawaiModel> getPegawaiInProvinsiById(BigInteger idProvinsi);
+	ProvinsiModel getProvinsiDetailById(long id);
+	List<PegawaiModel> getPegawaiInInstansiInProvinsiById(long idProvinsi, long idInstansi);
+	List<PegawaiModel> getPegawaiInProvinsiById(long idProvinsi);
 }
